Moved testing constants to test file
