+++ tests/instructions/to.test.ts
@@ -1,14 +1,13 @@
 import { expect, test } from 'bun:test';
 import { type Model, type Query, Transaction } from '@/src/index';
 
-import { queryEphemeralDatabase } from '@/fixtures/utils';
-import type { MultipleRecordResult, SingleRecordResult } from '@/src/types/result';
 import {
   RECORD_ID_REGEX,
   RECORD_TIMESTAMP_REGEX,
-  RONIN_MODEL_SYMBOLS,
-  RoninError,
-} from '@/src/utils/helpers';
+  queryEphemeralDatabase,
+} from '@/fixtures/utils';
+import type { MultipleRecordResult, SingleRecordResult } from '@/src/types/result';
+import { RONIN_MODEL_SYMBOLS, RoninError } from '@/src/utils/helpers';
 
 test('set single record to new string field', async () => {
   const queries: Array<Query> = [
